Combine image server code into a single class
